Refactoring sam code

+++ ja2lib/TileEngine/ExplosionControl.h
@@ -89,7 +89,7 @@ void SetOffPanicBombs(UINT8 ubID, INT8 bPanicTrigger);
 void UpdateExplosionFrame(INT32 iIndex, INT16 sCurrentFrame);
 void RemoveExplosionData(INT32 iIndex);
 
-void UpdateAndDamageSAMIfFound(u8 sSectorX, u8 sSectorY, INT16 sSectorZ, INT16 sGridNo,
+void UpdateAndDamageSAMIfFound(u8 sSectorX, u8 sSectorY, i8 sSectorZ, INT16 sGridNo,
                                UINT8 ubDamage);
 void UpdateSAMDoneRepair(u8 sSectorX, u8 sSectorY, INT16 sSectorZ);
